requirements.txt
# Core dependencies
faker>=20.0.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0

# Database
pyodbc>=5.0.0
pymssql>=2.2.0
sqlalchemy>=2.0.0

# Performance
multiprocessing-logging>=0.3.4
tqdm>=4.65.0
psutil>=5.9.0

# Data generation
phonenumbers>=8.13.0
email-validator>=2.0.0

# CLI
click>=8.1.0
rich>=13.5.0
tabulate>=0.9.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pre-commit>=3.3.0